% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{get_rules}
\alias{get_rules}
\title{List rules for creating new variables in \code{\link{load_data}}}
\usage{
get_rules()
}
\description{
List rules for creating new variables in \code{\link{load_data}}
}
