% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{load_data}
\alias{load_data}
\title{Load and pre-process data from relevant CHAMPS database tables}
\usage{
load_data(
  data_dir = NULL,
  dmg_path = NULL,
  lab_path = NULL,
  dcd_path = NULL,
  tac_path = NULL,
  rules = NULL
)
}
\arguments{
\item{data_dir}{Path to directory containing L3 csv files.
Note that if this is provided, the additional `path`
parameters are not required, but csv files must exist
with the naming scheme as described for these files.}

\item{dmg_path}{Path to L3 demographics csv file (typically
named "CHAMPS_L3_basic_demographics_V[x].csv").}

\item{lab_path}{Path to L3 lab results csv file (typically
named "CHAMPS_L3_lab_results_V[x].csv").}

\item{dcd_path}{Path to L3 DeCoDe results csv file (typically
named "CHAMPS_L3_decode_results_V[x].csv").}

\item{tac_path}{Path to L3 TAC results csv file (typically 
namde "CHAMPS_L3_tac_results_V[x].csv").}

\item{rules}{Optional list of rules for calculating new variables
(see \code{\link{get_rules}})}
}
\description{
Load and pre-process data from relevant CHAMPS database tables
}
\note{
The excel files must be fresh exports from the database not have had
  any processing, such as additional sheets added to them. One exception is
  that it is assumed that the TAC results file (indicated by
  \code{tac_res_path}) is assumed to have indeterminate results removed,
  although the logic can easily be changed so that this is not required. This
  function and all functions in this package rely on the input data coming
  directly from the database with no variable names changed, etc.
}
