% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_etiologies.R
\name{plot_etiologies}
\alias{plot_etiologies}
\title{Plot the etiologies of a computed CHAMPS table}
\usage{
plot_etiologies(x, etiologies = NULL, top = 6, bar_width = 0.5)
}
\arguments{
\item{x}{A computed CHAMPS object.}

\item{etiologies}{The etiologies to include in the graphic with their shortened names as a named vector. Defaults to NULL.  If NULL than the top variable is used and the names are shortened automatically.}

\item{top}{The number of etiologies to include in the graphic}

\item{bar_width}{defines the size of the bar the chart. Defaults to 0.5.}
}
\description{
Plot the etiologies of a computed CHAMPS table
}
