% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{plot_margins}
\alias{plot_margins}
\title{Plot the margins (row and column totals) of a computed CHAMPS table}
\usage{
plot_margins(x)
}
\arguments{
\item{x}{A computed CHAMPS object.}
}
\description{
Plot the margins (row and column totals) of a computed CHAMPS table
}
