% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_syndrome.R
\name{plot_syndrome}
\alias{plot_syndrome}
\title{Plot the syndrome distribution from a computed CHAMPS table}
\usage{
plot_syndrome(
  x,
  legend_location = "none",
  plot_type = "pie",
  include_pie_numbers = FALSE,
  include_bar_text = TRUE,
  waffle_cols = 12
)
}
\arguments{
\item{x}{A computed CHAMPS object.}

\item{legend_location}{Can be one of "none", "bottom", "right", "left". Defaults to none.}

\item{plot_type}{One of four optins. Pie Chart 'pie', Bar Chart 'bar', Treemap 'tree', Waffle 'waffle'}

\item{include_pie_numbers}{Whether to include the numbers within the slices of the pie chart. Defaults to FALSE.}

\item{include_bar_text}{Whether to include the category text within the bar. Defaults to TRUE}

\item{waffle_cols}{The number of columns of the waffle chart.}
}
\description{
Plot the syndrome distribution from a computed CHAMPS table
}
