% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{data}
\name{tac_table}
\alias{tac_table}
\title{tac_table}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 120 rows and 3 columns.}
\usage{
tac_table
}
\description{
tac_table
}
\keyword{datasets}
