% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{write_html_table}
\alias{write_html_table}
\title{Write an html table of a computed CHAMPS object}
\usage{
write_html_table(
  x,
  margin_pct = TRUE,
  margin_denom = FALSE,
  inside_pct = FALSE,
  inside_denom = FALSE
)
}
\arguments{
\item{x}{A computed CHAMPS object.}

\item{margin_pct}{Should percentages be printed in the margins?}

\item{margin_denom}{Should denominators be printed in the margins?}

\item{inside_pct}{Should percentages be printed inside the table?}

\item{inside_denom}{Should denominators be printed inside the table?}
}
\description{
Write an html table of a computed CHAMPS object
}
