% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_exp.R
\name{write_html_table_experimental}
\alias{write_html_table_experimental}
\title{Experimental html table}
\usage{
write_html_table_experimental(x)
}
\arguments{
\item{x}{Object of class "cc_allcases_by_site_casetype".}
}
\description{
Experimental html table
}
