% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{write_ppt_slide}
\alias{write_ppt_slide}
\title{Print a computed CHAMPS table to a PowerPoint presentation}
\usage{
write_ppt_slide(
  x,
  doc,
  margin_pct = TRUE,
  margin_denom = FALSE,
  inside_pct = FALSE,
  inside_denom = FALSE
)
}
\arguments{
\item{x}{A computed CHAMPS object.}

\item{doc}{A PowerPoint presentation object.}

\item{margin_pct}{Should percentages be printed in the margins?}

\item{margin_denom}{Should denominators be printed in the margins?}

\item{inside_pct}{Should percentages be printed inside the table?}

\item{inside_denom}{Should denominators be printed inside the table?}
}
\description{
Print a computed CHAMPS table to a PowerPoint presentation
}
